+++ src/gui/Coordinator.hpp
@@ -32,7 +32,8 @@ namespace mdl {
   class Coordinator
   {
   public:
-    Coordinator(FilterList& filter_list,
+    Coordinator(Gtk::Window& parent_window,
+                FilterList& filter_list,
                 FrameNavigator& frame_navigator,
                 int frame_width, int frame_height);
 
@@ -46,6 +47,8 @@ namespace mdl {
     int get_current_frame();
 
   private:
+    Gtk::Window& parent_window_;
+
     FilterList& filter_list_;
     Glib::RefPtr<FilterListModel> filter_model_;
 
@@ -55,6 +58,7 @@ namespace mdl {
 
     FilterPanelFactory panel_factory_;
     FilterPanel* current_filter_panel_;
+    int current_filter_start_frame_;
     fg::Filter* current_filter_;
     bool scroll_filter_;
 
@@ -76,6 +80,11 @@ namespace mdl {
     void on_frame_rectangle_changed(Rectangle rect);
     sigc::connection on_panel_rectangle_changed_;
     void on_panel_rectangle_changed(Rectangle rect);
+
+    sigc::connection on_start_frame_changed_;
+    void on_start_frame_changed(int start_frame);
+    bool confirm_overwrite_by_start_frame_change(int start_frame);
+    void set_start_frame_in_filter_panel(int start_frame);
 
     void create_new_filter_panel();
     void update_current_filter(bool force_updated);